fix: interface폴더 삭제 및 interface코드들 해당 컴포넌트로 이동
