+++ src/App.tsx
@@ -1,6 +1,5 @@
 /** @jsxImportSource @emotion/react */
 import { createBrowserRouter, Navigate, RouterProvider } from 'react-router-dom';
-import { IPrivateRouteProps } from '@/interfaces/privateRoute';
 import checkAuth from '@/api/auth/checkAuth';
 import Home from '@/pages/Home';
 import Schedule from '@/pages/Schedule/Schedule';
@@ -15,6 +14,10 @@ import CorrectionRequest from '@/pages/Wage/Correction/CorrectionRequest.tsx';
 import GlobalStyles from '@/styles/GlobalStyles.tsx';
 import Layout from '@/layout/Layout';
 import Login from '@/pages/Login';
+
+export interface IPrivateRouteProps {
+	element: JSX.Element;
+}
 
 // 로그인한 사용자만 접근할 수 있는 라우팅 페이지를 관리하기 위한 PrivateRoute
 const PrivateRoute = ({ element }: IPrivateRouteProps) => {
