+++ src/components/Calendar/CalendarDates.tsx
@@ -1,8 +1,12 @@
 import { mockdata, IMockDate } from '@/components/Calendar/mockdata';
-import { ICalendarDatesProps } from '@/interfaces/calendar';
 import { colors } from '@/constants/colors';
 import CalendarBadge from '@/components/Calendar/CalendarBadge';
 import styled from '@emotion/styled';
+
+export interface ICalendarDatesProps {
+	date: Date;
+	isOfficial: boolean;
+}
 
 const CalendarDates = ({ date, isOfficial }: ICalendarDatesProps) => {
 	// 날짜별 데이터 필터링
