+++ src/components/Input.tsx
@@ -1,6 +1,14 @@
-import { IInputProps } from '@/interfaces/input';
 import { colors } from '@/constants/colors';
 import styled from '@emotion/styled';
+
+export interface IInputProps {
+	value: string;
+	onChange: (event: React.ChangeEvent<HTMLInputElement>) => void;
+	inputWidth?: string;
+	placeholder?: string;
+	type?: 'text' | 'password' | 'date';
+	disabled?: boolean;
+}
 
 const Input = ({
 	value,
