--- src/interfaces/button.d.ts
@@ -1,8 +1,0 @@
-export interface IButtonProps {
-	label: string;
-	onClick: () => void;
-	size?: 'small' | 'normal';
-	theme?: 'primary' | 'secondary';
-	buttonWidth?: string;
-	disabled?: boolean;
-}
