--- src/interfaces/iconButton.d.ts
@@ -1,7 +1,0 @@
-export interface IIconButtonProps {
-	IconComponent: React.ComponentType;
-	shape: 'transparent' | 'line';
-	color?: 'gray' | 'black';
-	size?: number;
-	onClick?: () => void;
-}
