--- src/interfaces/menu.d.ts
@@ -1,6 +1,0 @@
-export interface IMenuItemProps {
-	to: string;
-	icon: JSX.Element;
-	label: string;
-	isActive: boolean;
-}
