--- src/interfaces/privateRoute.d.ts
@@ -1,3 +1,0 @@
-export interface IPrivateRouteProps {
-	element: JSX.Element;
-}
